upd (*) refactoring
